added some methods for NetworkProvider trait

+++ wallet/src/wallet_data.rs
@@ -47,7 +47,7 @@ pub struct WalletData {
     pub accounts: Vec<Account>,
     pub selected_account: usize,
     pub biometric_type: AuthMethod,
-    pub network: Vec<usize>,
+    pub network: Vec<u64>,
 }
 
 impl WalletData {
